@@ -3,25 +3,22 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.11.6",
-    "@testing-library/react": "^11.2.2",
-    "@testing-library/user-event": "^12.6.0",
-    "bootstrap": "^4.5.3",
-    "firebase": "^8.2.3",
-    "react": "^17.0.1",
-    "react-bootstrap": "^1.4.0",
-    "react-dom": "^17.0.1",
-    "react-icons": "^4.1.0",
-    "react-multi-carousel": "^2.6.1",
-    "react-redux": "^5.1.1",
-    "react-redux-firebase": "^2.2.4",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.1",
-    "redux": "^4.0.5",
-    "redux-firebase": "^0.3.3",
-    "redux-firestore": "^0.14.0",
-    "redux-thunk": "^2.3.0",
-    "web-vitals": "^0.2.4"
+    "@emotion/react": "^11.10.5",
+    "@mantine/core": "^5.10.0",
+    "@mantine/form": "^5.10.0",
+    "@mantine/hooks": "^5.10.0",
+    "@tabler/icons": "^1.119.0",
+    "axios": "^1.2.1",
+    "bootstrap": "^5.2.3",
+    "moment": "^2.29.4",
+    "react": "^18.2.0",
+    "react-bootstrap": "^2.7.0",
+    "react-dom": "^18.2.0",
+    "react-icons": "^4.7.1",
+    "react-multi-carousel": "^2.8.2",
+    "react-router-dom": "^6.5.0",
+    "react-scripts": "^5.0.1",
+    "react-toastify": "^9.1.1"
   },
   "scripts": {
     "start": "react-scripts start",
